refactor: add tracing package to consolidate
